+++ src/includes/traits/Facades/Markup.php
@@ -1,0 +1,41 @@
+<?php
+/**
+ * Markup.
+ *
+ * @author @jaswsinc
+ * @copyright WebSharks™
+ */
+declare(strict_types=1);
+namespace WebSharks\WpSharks\Core\Traits\Facades;
+
+use WebSharks\WpSharks\Core\Classes;
+use WebSharks\WpSharks\Core\Interfaces;
+use WebSharks\WpSharks\Core\Traits;
+#
+use WebSharks\Core\WpSharksCore\Classes as CoreClasses;
+use WebSharks\Core\WpSharksCore\Classes\Core\Base\Exception;
+use WebSharks\Core\WpSharksCore\Interfaces as CoreInterfaces;
+use WebSharks\Core\WpSharksCore\Traits as CoreTraits;
+#
+use function assert as debug;
+use function get_defined_vars as vars;
+
+/**
+ * Markup.
+ *
+ * @since 17xxxx
+ */
+trait Markup
+{
+    /**
+     * @since 17xxxx Markup utils.
+     *
+     * @param mixed ...$args Variadic args to underlying utility.
+     *
+     * @see Classes\SCore\Utils\Markup::errors()
+     */
+    public static function markupErrors(...$args)
+    {
+        return $GLOBALS[static::class]->Utils->§Markup->errors(...$args);
+    }
+}